refactor: move functions out of top init file
